requirements.txt
python-dotenv
jinja2
requests
reolink-aio
Pillow